[Frontend] Move Nickcompleter to Frontend code and improve it.
It is basically independent from any specific frontend, it just
needs an instance of IChatView, some text and position as the
input and can be applied to every frontend.

Furthermode I improved the completion, it will now iterate through
all the results, when not in bash style. AND it will sort the
the nicks according to people who last mentioned the target or
who last spoke.